+++ path_planning/src/rrt_node.h
@@ -12,19 +12,19 @@ class rrt_node {
 private:
     Point position;
     joint_angles angles;
-    vector<float> angles_;
+    vector<double> angles_;
     rrt_node* parent_node = nullptr;
     list<rrt_node*> children;
     rrt_params params;
 public:
     bool has_child = false;
     bool has_parent = true;
-    tuple<float, float> cost_two_joints(rrt_node* init, rrt_node* stepped);
+    tuple<double, double> cost_two_joints(rrt_node* init, rrt_node* stepped);
     rrt_node(const joint_angles angles, const Point position,rrt_params params);
     rrt_node(const joint_angles angles, const Point position,rrt_params params, rrt_node* parent_node);
-    float cost = 0;
-    float cost_to_parent = 0;
-    void set_parent(rrt_node* parent, float cost, float cost_to_parent){
+    double cost = 0;
+    double cost_to_parent = 0;
+    void set_parent(rrt_node* parent, double cost, double cost_to_parent){
         this->parent_node->remove_child(this);
         this->parent_node = parent;
         this->parent_node->add_child(this);
@@ -54,7 +54,7 @@ public:
     Point& get_position(){
         return position;
     };
-    vector<float>* get_angles_flann(){
+    vector<double>* get_angles_flann(){
         return &angles_;
     };
     rrt_node* get_parent(){
